backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Cloud
google-cloud-speech==2.23.0
google-generativeai==0.3.2

# Data handling
pandas==2.1.3

# File processing
ffmpeg-python==0.2.0
Pillow==10.1.0

# Utilities
aiofiles==23.2.1
